requirements.txt
caldav==1.3.9
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
icalendar==6.1.1
idna==3.10
lxml==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
recurring-ical-events==3.4.1
requests==2.32.3
six==1.17.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
urllib3==2.3.0
vobject==0.9.9
x-wr-timezone==2.0.0
